requirements.txt
D3DShot==0.1.3
opencv-python==4.1.2
Pillow==6.2.0
websockets==8.1
